started working on display library
