+++ OLED_displayer/Display.cpp
@@ -3,8 +3,78 @@
 
 Display::Display()
 {
+	/*
 	std::cout << "Opening connection I2C..." << std::endl;
 	_fd = open("/dev/i2c-1", O_RDWR);
-	if (fd < 0)
-		throw std::system_error("Error opening connection I2C");
+	if (_fd < 0)
+		throw (std::exception());
+
+	std::cout << "Configuring display address.." << std::endl;
+	if (ioctl(_fd, I2C_SLAVE, DISPLAY_ADDR) < 0)
+		throw (std::exception());
+	*/
+
+	_fd = STDOUT_FILENO; // TEST
+
+	_initDisplay();
 }
+
+Display::~Display()
+{
+	if (_fd > 0)
+		close(_fd);
+}
+
+void	Display::_initDisplay(void)
+{
+	_writeCmd(DISPLAY_OFF);
+	_writeCmd(SET_CONTRAST);
+	_writeCmd(0xFF);
+
+	_writeCmd(SET_MULTIPLEX_RATIO);
+	_writeCmd(31);
+
+	_writeCmd(SET_MEM_ADDR_MODE);
+	_writeCmd(HORIZONTAL);
+
+	_writeCmd(DISPLAY_START_LINE);
+	_writeCmd(SEG_REMAP_RIGHT);
+	_writeCmd(COM_SCAN_DOWN);
+
+	_writeCmd(COM_PINS_CONFIG);
+	_writeCmd(0x02);
+
+	_writeCmd(0xD5);
+	_writeCmd(0x80);
+
+	_writeCmd(0x8D);
+	_writeCmd(0x14);
+
+	_writeCmd(SET_PAGE_NO | 0);
+	_writeCmd(0x00);
+	_writeCmd(0x10);
+
+	_initBuffer();
+	updateDisplay();
+	
+	_writeCmd(NORMAL_DISPLAY);
+	_writeCmd(DISPLAY_ON);
+}
+
+void	Display::_initBuffer(void) {_buffer[0] = PXL_ADD;}
+
+void	Display::updateDisplay(void) const
+{
+	write(_fd, _buffer.data(), _buffer.size());
+}
+
+void	Display::_writeCmd(unsigned char data) const
+{
+	unsigned char	buffer[2];
+	buffer[0] = CMD_ADD;
+	buffer[1] = data;
+	write(_fd, buffer, 2);
+}
+
+void	Display::setPixel(int x, int y) {_buffer[y/8*WIDTH+x+1] |= 1<<(y%8);}
+void	Display::unsetPixel(int x, int y) {_buffer[y/8*WIDTH+x+1] &= ~(1<<(y%8));}
