--- OLED_displayer/test.cpp
@@ -1,6 +1,0 @@
-#include "Display.hpp"
-
-int	main(void)
-{
-//	Display	dp;
-}
